workaround: really big numbers in HTML input element
With an input of `type=number` Javascript will call the API end-point
with a 'e' notation (ex: user enters "6000000000000000000000", JS sends
{'measured': '6e21'}).
Reading the string back (page relad), nothing appears in the input
box because '6e21' is not considered a number.

Trying to solve the issue with
    measured = '{:.0f}'.format(float(measured))
leads to random digits being added on some machines (VMware Fusion /
Python 3.6.5). Using a `Decimal` seems to give constant results.